src/category/domain/category.entity.ts
import { Uuid } from "../../shared/domain/value-objects/uuid.vo"

export type CategoryConstructorProps = {
  category_id?: Uuid
  name: string
  description?: string | null
  is_active?: boolean
  created_at?: Date
}
export type CategoryCreateCommand = {
  name: string
  description?: string | null
  is_active?: boolean
}
export class Category {
  public category_id: Uuid
  public name: string
  public description: string | null
  public is_active: boolean
  public created_at: Date

  constructor (props: CategoryConstructorProps) {
    this.category_id = props.category_id ?? new Uuid()
    this.name = props.name
    this.description = props.description ?? null
    this.is_active = props.is_active ?? true
    this.created_at = props.created_at ?? new Date()
  }

  static create (props: CategoryCreateCommand): Category {
    return new Category(props)
  }

  public changeName (name: string): void {
    this.name = name
  }

  public changeDescription (description: string): void {
    this.description = description
  }

  public activate (): void {
    this.is_active = true
  }

  public deactivate (): void {
    this.is_active = false
  }

  public toJSON (): Object {
    return {
      category_id: this.category_id.id,
      name: this.name,
      description: this.description,
      is_active: this.is_active,
      created_at: this.created_at
    }
  }
}